API for authentication

+++ controllers/articles.js
@@ -6,7 +6,8 @@ exports.createArticle = (req, res) => {
     const article = new Article({
         title: req.body.title,
         content: req.body.content,
-        imageUrl
+        imageUrl,
+        userId: req.auth.userId 
     });
     article.save()
     .then(article => { res.status(200).json({ article })})
@@ -23,48 +24,67 @@ exports.getAllArticles = (req, res) => {
         .then(articles => res.status(200).json(articles))
         .catch(error => res.status(400).json({ error }));
 };
-
+ 
 exports.getOneArticle =  (req, res) => {
     Article.findOne({ _id: req.params.id })
         .then(article => res.status(200).json(article))
         .catch(error => res.status(404).json({ error }));
 }
 
- exports.deleteArticle = (req, res) => {
+exports.deleteArticle = (req, res) => {
     const articleId = req.params.id;
+    const userId = req.auth.userId; 
+
     Article.findOne({ _id: articleId })
         .then(article => {
             if (!article) {
                 return res.status(404).json({ message: 'Article non trouvé' });
             }
+            if (article.userId.toString() !== userId) {
+                return res.status(403).json({ message: 'Requête non autorisée' });
+            }
             const filename = article.imageUrl.split('/images/')[1];
             fs.unlink(`images/${filename}`, () => {
                 Article.deleteOne({ _id: articleId })
                     .then(() => res.status(200).json({ message: 'Article supprimé !' }))
-                    .catch(error => res.status(500).json({ error: error.message }));
             });
         })
-        .catch(error => res.status(500).json({ error: error.message }));
 };
 
 exports.modifyArticle = (req, res) => {
-    const update = {
-        title: req.body.title,
-        content: req.body.content
-    };
-    if (req.file) {
-        update.imageUrl = `${req.protocol}://${req.get('host')}/images/${req.file.filename}`;
-    }
-    Article.findOneAndUpdate(
-        { _id: req.params.id },
-        { ...update, _id: req.params.id },
-        { new: true }
-    )
-    .then((updatedArticle) => {
-        if (!updatedArticle) {
-            return res.status(404).json({ message: 'Article non trouvé' });
-        }
-        res.status(200).json({ article: updatedArticle });
-    })
-    .catch((error) => res.status(400).json({ error: error.message }));
+    const userId = req.auth.userId;
+
+    Article.findOne({ _id: req.params.id })
+        .then(article => {
+            if (!article) {
+                return res.status(404).json({ message: 'Article non trouvé' });
+            }
+            if (article.userId.toString() !== userId) {
+                return res.status(403).json({ message: 'Requête non autorisée' });
+            }
+
+            const update = {
+                title: req.body.title,
+                content: req.body.content
+            };
+            if (req.file) {
+                const oldFilename = article.imageUrl.split('/images/')[1];
+                fs.unlink(`images/${oldFilename}`, (err) => {
+                    if (err) console.error(err);
+                });
+                update.imageUrl = `${req.protocol}://${req.get('host')}/images/${req.file.filename}`;
+            }
+
+            Article.findOneAndUpdate(
+                { _id: req.params.id },
+                { ...update, _id: req.params.id },
+                { new: true }
+            )
+            .then((updatedArticle) => {
+                res.status(200).json({ article: updatedArticle });
+            })
+            .catch((error) => res.status(400).json({ error: error.message }));
+        })
+        .catch((error) => res.status(500).json({ error: error.message }));
 };
+
